requirements.txt
pandas==2.1.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.1.1
